#41 refactor : 차량 매물등록 플레이스홀더화 및 카메라 버튼 변경
